// ExistingCode/src/data-access/domain/tables/rigsetup/schema.ts
export const RigSetupTableInsertBaseSchema = z.object({
	/**
	 * dataType: uniqueidentifier
	 * defaultValue: (newid())
	 */
	rigSetupId: z.any().optional(),
	/**
	 * dataType: nvarchar
	 */
	organization: z.string().max(30),
	/**
	 * dataType: uniqueidentifier
	 */
	drillPlanId: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	padInspectionCompletedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	padInspectionSignature: z.string().max(-1).nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	padInspectionSignatureDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	drillingCompany: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	drillSupervisor: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	drillingSignature: z.string().max(-1).nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	drillingSignatureDt: z.any().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	finalMagAzimuth: z.number().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	finalInclination: z.number().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	finalGeologist: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	finalGeologistSignature: z.string().max(-1).nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	finalGeologistSignatureDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	finalSetupDrillSupervisor: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	finalSetupDrillSupervisorSignature: z.string().max(-1).nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	finalSetupDrillSupervisorSignatureDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	finalSetupApprovedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	finalSetupSignature: z.string().max(-1).nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	finalSetupSignatureDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	downHoleSurveyDrillingContractor: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	downHoleSurveyRigNo: z.string().max(50).nullable().optional(),
	/**
	 * dataType: decimal
	 */
	rigAlignmentToolMagAzi: z.number().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	rigAlignmentToolDip: z.number().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	surveyReference: z.string().max(50).nullable().optional(),
	/**
	 * dataType: decimal
	 */
	surveyMagAzi: z.number().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	surveyDip: z.number().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	surveyDepth: z.number().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	downHoleSurveyDriller: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	downHoleSurveyDrillerSignature: z.string().max(-1).nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	downHoleSurveyDrillerSignatureDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	comments: z.string().max(-1).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	dataSource: z.string().max(255),
	/**
	 * dataType: bit
	 * defaultValue: ((0))
	 */
	reportIncludeInd: z.any().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	validationStatus: z.number().int().optional(),
	/**
	 * dataType: nvarchar
	 */
	validationErrors: z.string().max(-1).nullable().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	rowStatus: z.number().int().optional(),
	/**
	 * dataType: uniqueidentifier
	 */
	supersededById: z.any().nullable().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((1))
	 */
	activeInd: z.any().optional(),
	/**
	 * dataType: datetimeoffset
	 * defaultValue: (sysdatetimeoffset())
	 */
	createdOnDt: z.any().optional(),
	/**
	 * dataType: nvarchar
	 * defaultValue: (CONVERT([nvarchar](50),coalesce(session_context(N'UserName'),suser_sname())))
	 */
	createdBy: z.string().max(50).optional(),
	/**
	 * dataType: datetimeoffset
	 */
	modifiedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	modifiedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: timestamp
	 */
	rv: z.date(),
});